Split witnessing workflow out of the countersigning workflow
