+++ crates/holochain/src/core/workflow.rs
@@ -34,6 +34,7 @@ pub mod integrate_dht_ops_workflow;
 pub mod publish_dht_ops_workflow;
 pub mod sys_validation_workflow;
 pub mod validation_receipt_workflow;
+pub mod witnessing_workflow;
 
 // MAYBE: either remove wildcards or add wildcards for all above child modules
 pub use call_zome_workflow::*;
